Added action and fixes serial handling
This commit adds the 'edit' action for domains. Also the 'name' attribute of domains and records are now read-only in edit view.

The change date of the soa is only updated if it's present. This removes a bug, where the system trys to update the change date of the SOA record, when the whole domain is deleted.